Add formula to install confluent-oss package
Create a formula to download and install the current (4.0.0)
version of the confluent oss package.

This was largely inspired by https://github.com/sgerrand/homebrew-confluent-platform
only that formula did not work for us, because at least as of 4/17/2018
that formula does not install the `share` directory with all the jar
files.

We are creating a asr specific `tap` so we can control this dependency

Specifying version, otherwise homebrew gets confused and thinks
the version is 2.11

Loosening the java requirement. According to
https://docs.confluent.io/current/installation/installing_cp.html#system-requirements
as of 4/17/18 java 1.7 or greater is needed. Our macs are 1.8 but 9 is
also available. Brew does not seem to support `depends_on` with >= or
bundler style ~>, so leaving it open to any java since odds are it will
be ok.